rename dask module to dask_cluster module
